unstructured_genieml_0.1.3/pyproject.toml
[project]
name = "kaiya-unstructureddata-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb==0.5.13",
    "fastapi[standard]>=0.115.11",
    "langchain>=0.1.0",
    "langchain-openai>=0.3.9",
    "langchain-community>=0.0.10",
    "langgraph>=0.3.18",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.1",
    "sqlalchemy>=2.0.0",
    "psycopg>=3.2.6",
    "psycopg2-binary>=2.9.0",
    "dagster >= 1.5.0",
    "dagit >= 1.5.0",
    "unstructured>=0.17.2",
    "unstructured-inference>=0.7.12",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "pi-heif>=0.13.1",
    "redis>=5.0.0",
    "unstructured-pytesseract>=0.3.10",
    "spacy>=3.8.7",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.0",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "langchain-google-genai==2.0.11",
    "PyPDF2==3.0.1"
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
]

[tool.setuptools]
py-modules = []

[tool.ruff.format]
exclude = ["*.ipynb"]

[tool.ruff.lint]
exclude = ["*.ipynb"]
extend-select = ["I"]
